+++ src/common/__tests__/migrationForMultiOptions.test.ts
@@ -9,17 +9,19 @@ import { EVersionedSchemaType } from '../VersionedSchema/Enums';
 import { migrateToLatestVersion } from '../VersionedSchema/Utils/migrateToLatestVersion';
 
 describe('migrateToLatestVersion for Multioptions', () => {
-    test.skip('@unit: should migrate Multioptions data from 1.0.0 to 2.0.0', () => {
+    test('@unit: should migrate Multioptions data from 1.0.0 to 2.0.0', () => {
         const dataMultiV1 = {
             input: 'source/path',
             output: 'generated/path',
             client: HttpClient.AXIOS,
-            items: [{ input: 'item1',output: 'generated/path', }, { input: 'item2',output: 'generated/path', }],
+            items: [{ input: 'source/path' }, { input: 'source/path' }],
         };
         const expectedDataMultiV3 = {
-            input: 'source/path',
             httpClient: HttpClient.AXIOS,
-            items: [{ input: 'item1' }, { input: 'item2' }],
+            items: [
+                { input: 'source/path', output: 'generated/path', outputCore: undefined, outputServices: undefined, outputModels: undefined, outputSchemas: undefined },
+                { input: 'source/path', output: 'generated/path', outputCore: undefined, outputServices: undefined, outputModels: undefined, outputSchemas: undefined },
+            ],
             useCancelableRequest: false,
         };
 
@@ -32,7 +34,7 @@ describe('migrateToLatestVersion for Multioptions', () => {
         assert.deepEqual(result, { value: expectedDataMultiV3, schemaVersion: '2.0.0', schemaType: EVersionedSchemaType.MULTI_OPTIONS }, 'Should migrate Multioptions V1 data to V3 correctly');
     });
 
-    test.skip('@unit: should migrate Multioptions data from 1.0.1 to 2.0.0', () => {
+    test('@unit: should migrate Multioptions data from 1.0.1 to 2.0.0', () => {
         const dataMultiV2 = {
             input: 'source/path',
             output: 'generated/path',
@@ -41,7 +43,10 @@ describe('migrateToLatestVersion for Multioptions', () => {
         };
         const expectedDataMultiV3 = {
             httpClient: HttpClient.FETCH,
-            items: [{ input: 'item1', output: 'generated/path', }, { input: 'item2', output: 'generated/path' }],
+            items: [
+                { input: 'source/path', output: 'generated/path', outputCore: undefined, outputServices: undefined, outputModels: undefined, outputSchemas: undefined },
+                { input: 'source/path', output: 'generated/path', outputCore: undefined, outputServices: undefined, outputModels: undefined, outputSchemas: undefined },
+            ],
         };
 
         const result = migrateToLatestVersion({
@@ -50,26 +55,30 @@ describe('migrateToLatestVersion for Multioptions', () => {
             migrationPlans: multiOptionsMigrationPlan,
         });
 
-        assert.deepEqual(result, { value: expectedDataMultiV3, schemaVersion: '2.0.0', schemaType: EVersionedSchemaType.MULTI_OPTIONS }, 'Should migrate Multioptions V2 data to V3 correctly');
+        assert.deepEqual(result, { value: expectedDataMultiV3, schemaVersion: '2.0.0', schemaType: EVersionedSchemaType.MULTI_OPTIONS }, 'Should migrate Multioptions 1.0.1 data to 2.0.0 correctly');
     });
 
-    test.skip('@unit: should return null for invalid Multioptions data after migration', () => {
+    test('@unit: should throw error if not conform any known version schema', () => {
         const dataMultiV1 = {
             input: 'source/path',
             output: 'generated/path',
-            client: 'invalid-client', // Некорректное значение
+            client: 'invalid-client',
             items: [{ input: 'item1' }, { input: 'item2' }],
         };
-        const result = migrateToLatestVersion({
-            rawInput: dataMultiV1,
-            versionedSchemas: multiOptionsVersionedSchema,
-            migrationPlans: multiOptionsMigrationPlan,
-        });
-
-        assert.equal(result, null, 'Should return null for invalid Multioptions data');
+        assert.throws(
+            () => {
+                migrateToLatestVersion({
+                    rawInput: dataMultiV1,
+                    versionedSchemas: multiOptionsVersionedSchema,
+                    migrationPlans: multiOptionsMigrationPlan,
+                });
+            },
+            /Data does not conform to any known version schema/,
+            'Should throw error if not conform any known version schema'
+        );
     });
 
-    test.skip('@unit: should throw error if no migration plan exists for Multioptions', () => {
+    test('@unit: should throw error if no migration plan exists for Multioptions', () => {
         const dataMultiV1 = {
             input: 'source/path',
             output: 'generated/path',